[Fix] removed "antd" imports from userTable controller
